refactor(core/update): tighten typing
- tighten typing across core/update by using Awaitable, Any, and more precise imports
- convert dict[str, object] hints to dict[str, Any] and guard version extraction
- adopt GlobalConfig, BackupService, PostDownloadProcessor, and ProgressReporter in signatures
- improve error handling in update_multiple_apps for cancelled/ex failed tasks
