+++ tests/core/verification/test_helpers.py
@@ -11,7 +11,7 @@ from __future__ import annotations
 import pytest
 
 from my_unicorn.core.api import Asset
-from my_unicorn.core.verification.helpers import (
+from my_unicorn.core.verify import (
     build_checksum_url,
     resolve_manual_checksum_file,
 )
